requirements.txt
keyboard==0.13.5
Pillow==8.0.1
pygame==1.9.6
